moved the chat message handling in player_observer to it's own function to be triggered by the main loop
I'm triggering the chat-message handler in the player_observer from the main loop now to address Issue #6. I'm still not happy with the telnet stuff at all, but this seems to make it a bit better. The telnet timeout is now dependent on the entire loops polling frequency

the chat command actions will still use the player_observers telnet instead of the main one, so i hope this is sufficient for now, until a more genius guy than i comes up with a brilliant idea!

This is actually the solution i dreamed about last night, i just forgot :)