integrated data-storage on a json basis. all stored data should be readable and adjustable this way. should be easy enough to use the data for external tools
the lobby is saved, players homes are saved
a players authentification status should be saved.

all data is stored according to the bots database prefix

i had to alter quite a lot and also introduce a few dirty 'hacks', so the code most likely needs to be cleaned up.
extensive testing will be needed!